+++ changelog.md
@@ -1,4 +1,5 @@
 # Future
+- [FEATURE] Add `isSoftDeleted` helper method to model instances [#7531](https://github.com/sequelize/sequelize/pull/7531)
 - [FEATURE] `addConstraint`, `removeConstraint`, `showConstraint` [#7108](https://github.com/sequelize/sequelize/pull/7108)
 - [FIXED] `changeColumn` generates incorrect query with ENUM type [#7455](https://github.com/sequelize/sequelize/pull/7455)
 - [ADDED] `options.alter` to sequelize.sync() to alter existing tables.[#537](https://github.com/sequelize/sequelize/issues/537)